requirements.txt
google-cloud-discoveryengine
google-api-core
google-auth
python-dotenv
streamlit